+++ config/internal/configsource/manager.go
@@ -186,20 +186,14 @@ func NewManager(_ *configparser.ConfigMap) (*Manager, error) {
 // in the configuration, returning a config.Parser fully resolved. This must be called only
 // once per lifetime of a Manager object.
 func (m *Manager) Resolve(ctx context.Context, parser *configparser.ConfigMap) (*configparser.ConfigMap, error) {
-	res := configparser.NewParser()
+	res := configparser.NewConfigMap()
 	allKeys := parser.AllKeys()
 	for _, k := range allKeys {
 		value, err := m.expandStringValues(ctx, parser.Get(k))
 		if err != nil {
-			// Call RetrieveEnd for all sources used so far but don't record any errors.
-			_ = m.retrieveEnd(ctx)
 			return nil, err
 		}
 		res.Set(k, value)
-	}
-
-	if errs := m.retrieveEnd(ctx); len(errs) > 0 {
-		return nil, consumererror.Combine(errs)
 	}
 
 	return res, nil
@@ -273,16 +267,6 @@ func (m *Manager) Close(ctx context.Context) error {
 	m.watchersWG.Wait()
 
 	return consumererror.Combine(errs)
-}
-
-func (m *Manager) retrieveEnd(ctx context.Context) []error {
-	var errs []error
-	for _, source := range m.configSources {
-		if err := source.RetrieveEnd(ctx); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	return errs
 }
 
 func (m *Manager) expandStringValues(ctx context.Context, value interface{}) (interface{}, error) {
@@ -484,7 +468,7 @@ func parseCfgSrc(s string) (cfgSrcName, selector string, params interface{}, err
 
 		if len(parts) > 1 && len(parts[1]) > 0 {
 			var cp *configparser.ConfigMap
-			cp, err = configparser.NewParserFromBuffer(bytes.NewReader([]byte(parts[1])))
+			cp, err = configparser.NewConfigMapFromBuffer(bytes.NewReader([]byte(parts[1])))
 			if err != nil {
 				return
 			}