+++ exporter/otlpexporter/otlp.go
@@ -55,7 +55,7 @@ func newExporter(cfg config.Exporter) (*exporter, error) {
 // start actually creates the gRPC connection. The client construction is deferred till this point as this
 // is the only place we get hold of Extensions which are required to construct auth round tripper.
 func (e *exporter) start(_ context.Context, host component.Host) (err error) {
-	e.w, err = newGrpcSender(e.config, host.GetExtensions())
+	e.w, err = newGrpcSender(e.config, host)
 	return
 }
 
@@ -94,8 +94,8 @@ type grpcSender struct {
 	callOptions    []grpc.CallOption
 }
 
-func newGrpcSender(config *Config, ext map[config.ComponentID]component.Extension) (*grpcSender, error) {
-	dialOpts, err := config.GRPCClientSettings.ToDialOptions(ext)
+func newGrpcSender(config *Config, host component.Host) (*grpcSender, error) {
+	dialOpts, err := config.GRPCClientSettings.ToDialOptions(host)
 	if err != nil {
 		return nil, err
 	}
@@ -123,17 +123,23 @@ func (gs *grpcSender) stop() error {
 }
 
 func (gs *grpcSender) exportTrace(ctx context.Context, td pdata.Traces) error {
-	_, err := gs.traceExporter.Export(gs.enhanceContext(ctx), td, gs.callOptions...)
+	req := otlpgrpc.NewTracesRequest()
+	req.SetTraces(td)
+	_, err := gs.traceExporter.Export(gs.enhanceContext(ctx), req, gs.callOptions...)
 	return processError(err)
 }
 
 func (gs *grpcSender) exportMetrics(ctx context.Context, md pdata.Metrics) error {
-	_, err := gs.metricExporter.Export(gs.enhanceContext(ctx), md, gs.callOptions...)
+	req := otlpgrpc.NewMetricsRequest()
+	req.SetMetrics(md)
+	_, err := gs.metricExporter.Export(gs.enhanceContext(ctx), req, gs.callOptions...)
 	return processError(err)
 }
 
 func (gs *grpcSender) exportLogs(ctx context.Context, ld pdata.Logs) error {
-	_, err := gs.logExporter.Export(gs.enhanceContext(ctx), ld, gs.callOptions...)
+	req := otlpgrpc.NewLogsRequest()
+	req.SetLogs(ld)
+	_, err := gs.logExporter.Export(gs.enhanceContext(ctx), req, gs.callOptions...)
 	return processError(err)
 }